@@ -10,8 +10,8 @@ use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
 /**
  * @ORM\Table("user")
  * @ORM\Entity
- * @UniqueEntity("email")
  */
+#[UniqueEntity('email')]
 class User implements UserInterface
 {
     /**
@@ -20,71 +20,63 @@ class User implements UserInterface
      * @ORM\GeneratedValue(strategy="AUTO")
      */
     private $id;
-
     /**
      * @ORM\Column(type="string", length=25, unique=true)
-     * @Assert\NotBlank(message="Vous devez saisir un nom d'utilisateur.")
      */
+    #[Assert\NotBlank(message: "Vous devez saisir un nom d'utilisateur.")]
     private $username;
-
     /**
      * @ORM\Column(type="string", length=64)
      */
     private $password;
-
     /**
      * @ORM\Column(type="string", length=60, unique=true)
-     * @Assert\NotBlank(message="Vous devez saisir une adresse email.")
-     * @Assert\Email(message="Le format de l'adresse n'est pas correcte.")
      */
+    #[Assert\NotBlank(message: 'Vous devez saisir une adresse email.')]
+    #[Assert\Email(message: "Le format de l'adresse n'est pas correcte.")]
     private $email;
-
     public function getId()
     {
         return $this->id;
     }
-
-    public function getUsername()
+    public function getUserIdentifier(): string
     {
         return $this->username;
     }
-
     public function setUsername($username)
     {
         $this->username = $username;
     }
-
     public function getSalt()
     {
         return null;
     }
-
     public function getPassword()
     {
         return $this->password;
     }
-
     public function setPassword($password)
     {
         $this->password = $password;
     }
-
     public function getEmail()
     {
         return $this->email;
     }
-
     public function setEmail($email)
     {
         $this->email = $email;
     }
-
-    public function getRoles()
+    public function getRoles(): array
     {
         return array('ROLE_USER');
     }
-
     public function eraseCredentials()
     {
+    }
+
+    public function getUsername()
+    {
+        // TODO: Implement getUsername() method.
     }
 }
